Address PR review polish: split COM value strategy, test empty-string, clarify hint
M3 (arch): move ComValueSetStrategy out of UiAutomationService.cs into a new
partial file UiAutomationService.ValueSetStrategy.cs, keeping the COM/marshalling
mechanics separate from the rest of the service and from the pure ValueSetter.

M2 (test): add empty-string coverage to ValueSetterTests — verify that clearing a
field ("") flows through ValuePattern unchanged, and that a non-numeric ""
skips RangeValuePattern and reaches LegacyIAccessible intact.

M4 (cli-ux): reword the set-value failure hint so the send-keys example leads
with '--via send-input' and explains that WinUI 3 / WPF rich text controls need
it (real keystrokes, foreground + unlocked desktop), while the post-message
default is silently dropped by the XAML pipeline but works for classic Win32.
