+++ src/winapp-CLI/WinApp.Cli/Services/UiAutomationService.cs
@@ -795,77 +795,6 @@ return Task.FromResult<UiElement?>(null);
         return Task.CompletedTask;
     }
 
-    /// <summary>
-    /// <see cref="IValueSetStrategy"/> backed by a live UI Automation COM element. Each method
-    /// acquires the relevant UIA pattern and performs the set, returning <c>false</c> (and logging at
-    /// debug level) when the pattern is unavailable or the COM call throws so the caller falls through
-    /// to the next mechanism.
-    /// </summary>
-    private sealed class ComValueSetStrategy(IUIAutomationElement comElement, ILogger logger) : IValueSetStrategy
-    {
-        public bool TrySetViaValuePattern(string text)
-        {
-            try
-            {
-                var pattern = (IUIAutomationValuePattern)comElement.GetCurrentPattern(UIA_PATTERN_ID.UIA_ValuePatternId);
-                unsafe
-                {
-                    var bstrPtr = Marshal.StringToBSTR(text);
-                    try
-                    {
-                        pattern.SetValue(new Windows.Win32.Foundation.BSTR((char*)bstrPtr));
-                    }
-                    finally
-                    {
-                        Marshal.FreeBSTR(bstrPtr);
-                    }
-                }
-                return true;
-            }
-            catch (Exception ex)
-            {
-                logger.LogDebug("ValuePattern.SetValue failed, trying fallbacks: {Message}", ex.Message);
-                return false;
-            }
-        }
-
-        public bool TrySetViaRangeValuePattern(double value)
-        {
-            try
-            {
-                var rangePattern = (IUIAutomationRangeValuePattern)comElement.GetCurrentPattern(UIA_PATTERN_ID.UIA_RangeValuePatternId);
-                rangePattern.SetValue(value);
-                return true;
-            }
-            catch (Exception ex)
-            {
-                logger.LogDebug("RangeValuePattern.SetValue failed: {Message}", ex.Message);
-                return false;
-            }
-        }
-
-        public bool TrySetViaLegacyIAccessible(string text)
-        {
-            try
-            {
-                var legacyPattern = (IUIAutomationLegacyIAccessiblePattern)comElement.GetCurrentPattern(UIA_PATTERN_ID.UIA_LegacyIAccessiblePatternId);
-                unsafe
-                {
-                    fixed (char* valuePtr = text)
-                    {
-                        legacyPattern.SetValue(new Windows.Win32.Foundation.PCWSTR(valuePtr));
-                    }
-                }
-                return true;
-            }
-            catch (Exception ex)
-            {
-                logger.LogDebug("LegacyIAccessible.SetValue failed: {Message}", ex.Message);
-                return false;
-            }
-        }
-    }
-
     public Task FocusAsync(UiSessionInfo session, UiElement element, CancellationToken ct)
     {
         _logger.LogDebug("Focusing element {ElementId}", element.Id);
